--- src/test/java/tests/mobile/TestBaseLocaleDrivers.java
@@ -1,37 +1,0 @@
-package tests.mobile;
-
-import com.codeborne.selenide.Configuration;
-import drivers.BrowserstackDriver;
-import drivers.LocalDriver;
-import helpers.Attach;
-import io.qameta.allure.selenide.AllureSelenide;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-
-import static com.codeborne.selenide.Selenide.*;
-import static com.codeborne.selenide.logevents.SelenideLogger.addListener;
-
-public class TestBaseLocaleDrivers {
-    static String deviceHost = System.getProperty("deviceHost");
-    @BeforeAll
-    public static void setup() {
-        addListener("AllureSelenide", new AllureSelenide());
-
-        Configuration.browser = LocalDriver.class.getName();
-        Configuration.browserSize = null;
-    }
-
-    @BeforeEach
-    public void startDriver() {
-        addListener("AllureSelenide", new AllureSelenide());
-        open();
-    }
-
-    @AfterEach
-    public void afterEach() {
-        String sessionId = sessionId().toString();
-        closeWebDriver();
-        Attach.addVideo(sessionId);
-    }
-}
